Access developer update messages from database
The bot now retrieves developer update messages from a table within the
database. Previously, these messages were hardcoded as a String inside
the code, which meant that the bot had to be redeployed everytime that
the developer wanted to change the message. This is now done by
inserting the message into the database table.

Additional log statements have also been inserted such that each update
received by the bot will minimally trigger one log message. This allows
for easier testing and bug fixing of the bot.