Change CallbackData format
CallbackData is only capable of holding a String of 64 bytes, which
was too small for some of the strings written using the old CallbackData
format. The format has now been changed to:
"location;locationId;stall;stallId"
which drastically reduces the size of the string and prevents most, if
not all, cases of the string exceeding the maximum size.

The code was also updated to handle the new format of CallbackData.