Add halal status feature
Known halal stalls will now be indicated in the list of stalls, and the
halal status of each stall will be displayed in the photo caption. This
is to allow Muslim users to quickly see if there are any halal stalls in
the location that they have chosen.

The code for building the caption has also been refactored and packaged
into a separate function to keep the code clean.